+++ src/integrity/errors.js
@@ -1,0 +1,8 @@
+export class IntegrityError extends Error {
+  constructor(code, message, details = {}) {
+    super(message);
+    this.name = 'IntegrityError';
+    this.code = code;
+    this.details = details;
+  }
+}